Se realizaron excepciones personalizadas para el manejo de Relaciones en entidades
# Conflicts:
#	src/main/java/org/code_wizards/Sistema_Budget/web/exception/RestExceptionHandlerGasto.java

+++ src/main/java/org/code_wizards/Sistema_Budget/dominio/dto/IngresosDto.java
@@ -10,7 +10,7 @@ import java.time.LocalDate;
 public record IngresosDto(
         Long idIngreso,
         @NotNull(message = "El id del presupuesto es obligatorio")
-        Integer idBudget,
+        Long idBudget,
         @NotNull(message = "El id de la categoria es obligatorio")
         Integer idCategory,
         @NotBlank(message = "La descripcion del ingreso es obligatoria")
