+++ src/main/java/org/code_wizards/Sistema_Budget/dominio/exception/IngresosYaExisteException.java
@@ -2,7 +2,7 @@ package org.code_wizards.Sistema_Budget.dominio.exception;
 
 public class IngresosYaExisteException extends RuntimeException {
 
-    public IngresosYaExisteException(Integer idBudget, String descriptionEntry) {
+    public IngresosYaExisteException(Long idBudget, String descriptionEntry) {
         super("El ingreso con Presupuesto ID: " + idBudget + " y Descripción: '" + descriptionEntry + "' ya existe.");
     }
 }
